requirements.txt
openai==0.27.8
discord.py==2.2.2
wolframalpha==5.0.0
requests==2.28.2
colorama==0.4.6
trafilatura==1.4.1
yt-dlp
aiodns==3.0.0
aiohttp==3.8.4
python-dotenv
